Renamed include file for channel types
