+++ src/resonanceReconstruction/rmatrix/ChannelType.hpp
